server/requirements.txt
fastapi
uvicorn
transformers
qwen-vl-utils
pillow
torch
